each user has their own cart stored on firebase, debug auto-logout, reduced it to 30 mins, fixed UI

+++ src/Components/Products/ProductItem/ProductItem.tsx
@@ -3,9 +3,10 @@ import { Product } from "../../../types";
 import Card from "../../UI/Card/Card";
 import { useDispatch } from "react-redux";
 import { AppDispatch } from "../../../store/redux-store";
-import { cartActions } from "../../../store/cart-slice";
 import Button from "../../UI/Button/Button";
 import { uiActions } from "../../../store/ui-slice";
+import { addToCart } from "../../../store/cart-thunks";
+import { getToken } from "../../../Util/token";
 
 interface Props {
   product: Product;
@@ -13,18 +14,22 @@ interface Props {
 }
 
 export default function ProductItem(props: Props) {
+  const uid = getToken();
   const dispatch = useDispatch<AppDispatch>();
   const addtoCartHandler = () => {
-    dispatch(
-      uiActions.addNotification({
-        title: "item added to cart!",
-        type: "success",
-      })
-    );
-    setTimeout(() => {
-      dispatch(uiActions.removeNotification());
-    }, 1500);
-    dispatch(cartActions.addItem({ ...props.product, amount: 1 }));
+    if (uid) {
+      dispatch(addToCart(uid, { ...props.product, amount: 1 }));
+    } else {
+      dispatch(
+        uiActions.addNotification({
+          title: "Please Login to add items to Cart!",
+          type: "error",
+        })
+      );
+      setTimeout(() => {
+        dispatch(uiActions.removeNotification());
+      }, 2000);
+    }
   };
   return (
     <Card className={classes.productContainer}>
